+++ packages/patrol_cli/lib/src/commands/build_ios.dart
@@ -171,6 +171,7 @@ class BuildIOSCommand extends PatrolCommand {
       appServerPort: super.appServerPort,
       testServerPort: super.testServerPort,
       clearPermissions: boolArg('clear-permissions'),
+      noTreeShakeIcons: boolArg('no-tree-shake-icons'),
     );
 
     try {